+++ src/mocks/handlers/index.ts
@@ -1,0 +1,6 @@
+import { feedHandlers } from './feed';
+
+/**
+ * feed.ts에서 작성한 feedHandlers를 handlers 배열에 삽입합니다.
+ */
+export const handlers = [...feedHandlers];
